+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # Change Log
-All notable changes to the "Advanced Tools Suite for AL Language" extension will be documented in this file.
 
 ## [1.0.0]
 
@@ -7,4 +6,10 @@ Initial release of "Advanced Tools Suite for AL Language"
 
 ## [1.2.0]
 
-Diagnostics: Namespace validation rules
+- New Features:
+    - Diagnostics: Namespace validation rules
+
+## [1.3.0]
+
+- New Features:
+    - `URL Forwarding Rules` used by command `ATS: Run Business Central without publishing` to run Business Central Client with different URL instead of the URL defined into launch.json for each configuration (Public URL or http to https conversion).